feat: Add new icons and update dependencies
This commit adds new icons for Back, Cart, Bell, LoveOutline, and Love
in the IconButton component. It also updates the project's dependencies,
including @react-navigation/stack,
react-native-animated-pagination-dots, react-native-shared-element, and
react-native-shimmer-placeholder.

These changes enhance the visual appeal and functionality of the app,
improving the user's experience. The updated dependencies ensure
compatibility with the latest versions and provide access to new
features and bug fixes.