Refactor app initialization and improve user experience (#123)
Simplify App.js imports and enhance the app startup process by
integrating the react-native-bootsplash library. This change ensures a
smoother transition from the splash screen to the main content.
Additionally, it hides the splash screen once the navigation container
is ready, providing a more polished user experience. The modifications
in the AndroidManifest.xml and styles.xml files are to support the
custom boot splash screen. This commit also includes changes to the
MainActivity.java file to initialize the boot splash theme. The changes
in index.js enable handling background messages and updating the unread
notification status. Lastly, the required storyboard and framework files
are added to the Xcode project. These changes improve the app's
performance, aesthetics, and usability.